Mostrar o cardápio de um país específico, Mostrar o cardápio de todos os países, Adicionar item ao cardápio, Remover Item do Cardápio, Mostrar a lista de restaurantes de um país

+++ NecTronalds/NecTronalds/Cardapio.cs
@@ -15,7 +15,7 @@ namespace NecTronalds
         {
             this.Lanches.Add(l);
         }
-        public void AdicionarLanche(string Nome, string Desc, float Preco)
+        public void AdicionarLanche(string Nome, string Desc, decimal Preco)
         {
             this.AdicionarLanche(new Lanche(Nome, Desc, Preco));
         }
